Card sheets are ready for front and back duplex printing. Structure improvement.
